// apps/api/src/modules/shops/dto/create-shop.input.d.ts
export declare class CreateShopInput {
    name: string;
    description?: string;
    phone?: string;
    country: string;
    timezone: string;
    currency: string;
}
export declare class UpdateShopInput {
    name?: string;
    description?: string;
    logoUrl?: string;
    coverUrl?: string;
}